+++ src/GUI/Categories/SearchNode.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../../Client/CategoryNode.hpp"
+#include <ColourModule.hpp>
 #include "../BetterInputNode.hpp"
 
 class SearchShowOptions : public Module
@@ -13,6 +14,16 @@ class SearchShowOptions : public Module
         }
 
         virtual void onToggle();
+};
+
+class SearchOptionsColour : public ColourModule
+{
+    public:
+        MODULE_SETUP(SearchOptionsColour)
+        {
+            setID("search/options-colour");
+            setDefaultConfig({ ccc3(255, 175, 0) });
+        }
 };
 
 class SearchNode : public CategoryNode, public TextInputDelegate
@@ -31,6 +42,8 @@ class SearchNode : public CategoryNode, public TextInputDelegate
 
         virtual void textChanged(CCTextInputNode* input);
         virtual void addModule(Module* module);
+
+        void updateColour(float);
 
         void onJoinDiscord(CCObject* sender);
         void onFilter(CCObject* sender);