+++ changelog.md
@@ -1,3 +1,13 @@
+# 2.7.4
+
+- Added Pitch Shifter high quality
+- Added Paste button to labels format page
+- Added a background behind options in the search page
+- Made category scrolls to save between popup opens
+- Improved search tab by using fuzzy search
+- Fixed show hitboxes not updating when rotating a slope in the editor
+- Fixed image labels being invisible
+
 # 2.7.3
 
 - Added UI Scale